Removed retry handling from MMeowlink
MMeowlink was re-trying certain communications automatically. This would cause
issues in certain cases where the state of the command object and the state of
the pump 'drifted' due to a retransmit.

This would show up in one of two ways:

1. Occasional page CRC errors in pump history, since a 'middle page' would be
   missing.
2. Temporary Basals would be set 3 times, rather than only once.

This change removes all retry handling from the MMeowlink code, since MMeowlink
doesn't have enough knowledge of the protocol to retry in a sane way.

The MMeowlink code will now throw an error that needs to be retried at a
different layer. This could be at the decocare layer, or at the openaps layer
(or indeed in the shell script that tries to call openaps!)