[comms_detection] Added mmeowlink-any-pump-comms.py
So that we can have two devices attempting to communicate with the pump at any
one time, we now include a command mmeowlink-any-pump-comms.py that can check
to see if there are ongoing pump comms - and return status appropriately.

If the state is 0 then it means that there *ARE* pump comms, as per the unix
convention that means it returned successfully.

eg:
 mmeowlink-any-pump-comms.py --port /dev/ttyACM0 --wait-for 3 --ignore-wake && echo "no comms"

If there have not been any communications in 3 seconds, this would return '1',
which means the `echo "no comms"` command would execute.