[adjust_frequency_ranges] Change frequency range
Narrow down the EU/WW range, and increase the USA range.

Background:

The Medtronic site (URL below) says the frequency for the Guardian is 868.35 and
the USA pump is 916.5

Since the USA frequency is very well optimised, I've tried keeping a similar
size range to the USA frequency, starting close to the 868.35 value like the USA
range starts at 916.5.

There is at least one report of 916.5 being too low, so I've extended it on the
downside too.

I've also increased the number of increment to account for the wider range.
This will make the scan take a bit longer to complete.

http://www.medtronicdiabetes.com/sites/default/files/library/download-library/user-guides/guardian_real_time_user_guide.pdf